+++ docs/changelog.rst
@@ -16,9 +16,12 @@ This is a major refactoring to adopt Test Driven Development, cleaner behaviors 
 * Removed everything about the CodeMirror rst editor (may probably live again in its own app);
 * Removed deprecated templates;
 * Replaced old Codemirror git submodule with a static copy from ``5.18.2`` version;
+* Dropped ``mode`` as an internal parameter since CodeMirror can use it in many different ways (as a string for a name, as a string for a mime-type, as a dict of mode options);
 
 Todo:
 
-* Don't use 'mode' parameter anymore as internal parameter since 'mode' parameter can be used in various forms (a simple mode name, a mode mime-type, a dict of mode options);
-* Finish and validate basic configs for some modes;
+* Finish documentation
+
+  * Ensure everything is documented about config;
+
 * Added tox configuration;